fix: update this module to work with latest commitizen v2.8.1
BREAKING CHANGE:
you need to update commitizen to v2.8.1, which uses inquireJs v1.0.2. From now on commitizen uses
promise for the commit callback. As a user of this adapter, all you have to do is update commitizen.

ISSUES CLOSED: 30
