+++ spec/czCustomizableSpec.js
@@ -32,96 +32,56 @@ describe('cz-customizable', function() {
     commit = jasmine.createSpy();
   });
 
-  it('should escape special characters sush as backticks', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
+  function getMockedCz(answers) {
+    return {
+      prompt: function() {
+        return {
+          then: function (cb) {
+            cb(answers);
+          }
+        };
+      }
+    };
+  }
 
+  it('should commit without confirmation', function() {
+    var answers = {
+      confirmCommit: 'yes',
+      type: 'feat',
+      subject: 'do it all'
+    };
+
+    var mockCz = getMockedCz(answers);
+
+    // run commitizen plugin
+    module.prompter(mockCz, commit);
+
+    expect(commit).toHaveBeenCalledWith('feat: do it all');
+  });
+
+  it('should escape special characters sush as backticks', function() {
     var answers = {
       confirmCommit: 'yes',
       type: 'feat',
       subject: 'with backticks `here`'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
+
     expect(commit).toHaveBeenCalledWith('feat: with backticks \\\\`here\\\\`');
   });
 
-  it('should call cz.prompt with questions', function() {
-    module.prompter(cz, commit);
+  it('should not call commit() function if there is no final confirmation and display log message saying commit has been canceled', function() {
+    var mockCz = getMockedCz({});
 
-    var getQuestion = function(number) {
-      return cz.prompt.mostRecentCall.args[0][number - 1];
-    };
+    // run commitizen plugin
+    module.prompter(mockCz, commit);
 
-    //question 1
-    expect(getQuestion(1).name).toEqual('type');
-    expect(getQuestion(1).type).toEqual('list');
-    expect(getQuestion(1).choices[0]).toEqual({value: 'feat', name: 'feat: my feat'});
-
-    //question 2
-    expect(getQuestion(2).name).toEqual('scope');
-    expect(getQuestion(2).choices({})[0]).toEqual({name: 'myScope'});
-    expect(getQuestion(2).choices({type: 'fix'})[0]).toEqual({name: 'fixOverride'}); //should override scope
-    expect(getQuestion(2).when({type: 'fix'})).toEqual(true);
-    expect(getQuestion(2).when({type: 'WIP'})).toEqual(false);
-    expect(getQuestion(2).when({type: 'wip'})).toEqual(false);
-
-    //question 3
-    expect(getQuestion(3).name).toEqual('scope');
-    expect(getQuestion(3).when({scope: 'custom'})).toEqual(true);
-    expect(getQuestion(3).when({scope: false})).toEqual(false);
-    expect(getQuestion(3).when({scope: 'scope'})).toEqual(false);
-
-    //question 4
-    expect(getQuestion(4).name).toEqual('subject');
-    expect(getQuestion(4).type).toEqual('input');
-    expect(getQuestion(4).message).toMatch(/IMPERATIVE tense description/);
-    expect(getQuestion(4).validate()).toEqual(false); //mandatory question
-    expect(getQuestion(4).filter('Subject')).toEqual('subject');
-
-    //question 5
-    expect(getQuestion(5).name).toEqual('body');
-    expect(getQuestion(5).type).toEqual('input');
-
-    //question 6
-    expect(getQuestion(6).name).toEqual('breaking');
-    expect(getQuestion(6).type).toEqual('input');
-    expect(getQuestion(6).when({type: 'feat'})).toEqual(true);
-    expect(getQuestion(6).when({type: 'fix'})).toEqual(false);
-
-    //question 7
-    expect(getQuestion(7).name).toEqual('footer');
-    expect(getQuestion(7).type).toEqual('input');
-    expect(getQuestion(7).when({type: 'fix'})).toEqual(true);
-    expect(getQuestion(7).when({type: 'WIP'})).toEqual(false);
-
-    //question 8, last one
-    expect(getQuestion(8).name).toEqual('confirmCommit');
-    expect(getQuestion(8).type).toEqual('expand');
-
-
-    var answers = {
-      confirmCommit: 'yes',
-      type: 'feat',
-      scope: 'myScope',
-      subject: 'create a new cool feature'
-    };
-    expect(getQuestion(8).message(answers)).toMatch('Are you sure you want to proceed with the commit above?');
-  });
-
-  it('should not call commit() function if there is no final confirmation', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
-    var res = commitAnswers({});
-
-    expect(res).toEqual(undefined);
     expect(commit).not.toHaveBeenCalled();
   });
 
   it('should call commit() function with commit message when user confirms commit and split body when pipes are present', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
-
     var answers = {
       confirmCommit: 'yes',
       type: 'feat',
@@ -132,14 +92,13 @@ describe('cz-customizable', function() {
       footer: 'my footer'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
+
     expect(commit).toHaveBeenCalledWith('feat(myScope): create a new cool feature\n\n-line1\n-line2\n\nBREAKING CHANGE:\nbreaking\n\nISSUES CLOSED: my footer');
   });
 
   it('should call commit() function with commit message with the minimal required fields', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
-
     var answers = {
       confirmCommit: 'yes',
       type: 'feat',
@@ -147,27 +106,24 @@ describe('cz-customizable', function() {
       subject: 'create a new cool feature'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
     expect(commit).toHaveBeenCalledWith('feat(myScope): create a new cool feature');
   });
 
   it('should suppress scope when commit type is WIP', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
-
     var answers = {
       confirmCommit: 'yes',
       type: 'WIP',
       subject: 'this is my work-in-progress'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
     expect(commit).toHaveBeenCalledWith('WIP: this is my work-in-progress');
   });
 
   it('should allow edit message before commit', function(done) {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
     process.env.EDITOR = 'true';
 
     var answers = {
@@ -176,7 +132,9 @@ describe('cz-customizable', function() {
       subject: 'create a new cool feature'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
+
     setTimeout(function() {
       expect(commit).toHaveBeenCalledWith('feat: create a new cool feature');
       done();
@@ -184,8 +142,6 @@ describe('cz-customizable', function() {
   });
 
   it('should not commit if editor returned non-zero value', function(done) {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
     process.env.EDITOR = 'false';
 
     var answers = {
@@ -194,7 +150,9 @@ describe('cz-customizable', function() {
       subject: 'create a new cool feature'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
+
     setTimeout(function() {
       expect(commit.wasCalled).toEqual(false);
       done();
@@ -202,9 +160,6 @@ describe('cz-customizable', function() {
   });
 
   it('should truncate first line if number of characters is higher than 200', function() {
-    module.prompter(cz, commit);
-    var commitAnswers = cz.prompt.mostRecentCall.args[1];
-
     var chars_100 = '0123456789-0123456789-0123456789-0123456789-0123456789-0123456789-0123456789-0123456789-0123456789-0123456789';
 
     // this string will be prepend: "ISSUES CLOSED: " = 15 chars
@@ -219,7 +174,8 @@ describe('cz-customizable', function() {
       footer: footerChars_100 + ' footer-second-line'
     };
 
-    commitAnswers(answers);
+    var mockCz = getMockedCz(answers);
+    module.prompter(mockCz, commit);
 
     var firstPart = 'feat(myScope): ';
 
@@ -235,100 +191,5 @@ describe('cz-customizable', function() {
     expect(footer).toEqual('ISSUES CLOSED: ' + footerChars_100 + '\nfooter-second-line');
 
   });
-
-
-  describe('optional fixOverride and allowBreakingChanges', function() {
-
-    beforeEach(function() {
-      module.__set__({
-        readConfigFile: function() {
-          return {
-            types: [{value: 'feat', name: 'feat: my feat'}],
-            scopes: [{name: 'myScope'}]
-          };
-        }
-      });
-
-      cz = jasmine.createSpyObj('cz', ['prompt', 'Separator']);
-      commit = jasmine.createSpy();
-    });
-
-
-    it('should call cz.prompt with questions', function() {
-      module.prompter(cz, commit);
-
-      var getQuestion = function(number) {
-        return cz.prompt.mostRecentCall.args[0][number - 1];
-      };
-
-      //question 1
-      expect(getQuestion(1).name).toEqual('type');
-      expect(getQuestion(1).type).toEqual('list');
-      expect(getQuestion(1).choices[0]).toEqual({value: 'feat', name: 'feat: my feat'});
-
-      //question 2
-      expect(getQuestion(2).name).toEqual('scope');
-      expect(getQuestion(2).choices({})[0]).toEqual({name: 'myScope'});
-      expect(getQuestion(2).choices({type: 'fix'})[0]).toEqual({name: 'myScope'}); //should override scope
-      expect(getQuestion(2).when({type: 'fix'})).toEqual(true);
-      expect(getQuestion(2).when({type: 'WIP'})).toEqual(false);
-      expect(getQuestion(2).when({type: 'wip'})).toEqual(false);
-
-      //question 6
-      expect(getQuestion(6).name).toEqual('breaking');
-      expect(getQuestion(6).when({type: 'feat'})).toEqual(true);
-      expect(getQuestion(6).when({type: 'fix'})).toEqual(true);
-      expect(getQuestion(6).when({type: 'FIX'})).toEqual(true);
-
-      var answers = {
-        confirmCommit: 'yes',
-        type: 'feat',
-        scope: 'myScope',
-        subject: 'create a new cool feature'
-      };
-      expect(getQuestion(8).message(answers)).toMatch('Are you sure you want to proceed with the commit above?');
-    });
-  });
-
-  describe('optional scopes', function() {
-
-    beforeEach(function() {
-      module.__set__({
-        readConfigFile: function() {
-          return {
-            types: [{value: 'feat', name: 'feat: my feat'}],
-            scopeOverrides: {
-              feat: [{name: 'myScope'}]
-            }
-          };
-        }
-      });
-
-      cz = jasmine.createSpyObj('cz', ['prompt', 'Separator']);
-      commit = jasmine.createSpy();
-    });
-
-
-    it('should call cz.prompt with questions', function() {
-      module.prompter(cz, commit);
-
-      var getQuestion = function(number) {
-        return cz.prompt.mostRecentCall.args[0][number - 1];
-      };
-
-      //question 2
-      expect(getQuestion(2).name).toEqual('scope');
-      expect(getQuestion(2).choices({})[0]).toBeUndefined();
-      expect(getQuestion(2).choices({type: 'feat'})[0]).toEqual({name: 'myScope'}); //should override scope
-      expect(getQuestion(2).when({type: 'feat'})).toEqual(true);
-      (function () {
-        var answers = {type: 'fix'};
-        expect(getQuestion(2).when(answers)).toEqual(false);
-        expect(answers.scope).toEqual('custom');
-      })();
-
-    });
-  });
-
 
 });
